Adding alias_method tests for #call
While the method is defined in #call the specs are testing the alias
of #[].

For the projects I'm considering using dry-data, I find the #call idiom to be
more inline with the projects existing idioms. So I want to see specs that
assert #call is part of the public interface and expectation.