Coerce empty strings in form posts to blank arrays and hashes
This makes it possible for forms to post blank arrays and hashes when the form post corresponds to dry-validation schema keys that are typed as arrays and hashes. You can do this with blank hidden inputs in your form, e.g.

<input name="some_array" type="hidden">
<input name="some_hash" type="hidden">

These hidden inputs would be posted with the form as empty strings, which these new coercions will turn into proper empty arrays and hashes.

This is useful because there is otherwise no way to signify in a form post that an array or hash value should be empty.