Allow Hash descendants in Hash::Weak schemas
Taking the pragmatic path here - because rails users may have to deal
with ActiveSupport::HashWithIndifferentAccess objects, using Hash::Weak
schemas before this change would result in failures due to
`instance_of?` check, this commit reverts the behavior back to how it
used to be and uses `is_a?` which will accept a Hash or a Hash
descendant objects.

This is not-so-great due to the weird nature of HWIA, but given that
it's used by Rails, lots of people would have a problem using dry-types
(and effectively dry-validation, too) as it wouldn't work at all with
HWIA. In general subclassing core types like Hash is a bad idea, as it
includes caveats in the form of unexpected object behaviors[*].

Anyhow, folks who don't use Rails won't be affected by this in a
negative way, as most commonly they will use plain hashes, and folks who
do use Rails, will not report issues and waste time trying to figure out
why dry-types/validation don't work with their parameters. So, sounds
like a win-win.

[*] http://words.steveklabnik.com/beware-subclassing-ruby-core-classes

Closes dry-rb/dry-validation#271