Add Object type
fixes #123

It's a small useful addition allows you to have a placeholder
type that can be later constrained with an actual type check.
It's also not so uncommon in a dynamically typed language to be
in a situation where you just don't know beforehand an object of
what type you're going to have. For instance, SQLite despite
being a SQL database can return you data of any type it supports
in any field of any row, and we have to deal with it.