Do not require value for default type
This also changes the existing behavior regarding working with `nil`.
From now on when a default type gets `nil` it tries to pass it through
the wrapped type first and see if the output is `nil` or not. If it's
`nil` then the default value will be returned. The reason behind this change
is to make hash schemas simpler in presence of default values. The goal is
to always call call a default type in a hash if the key is missing.
The behavior in presence of `nil` was dictated by backward compatibility,
we _may_ change it in future once it causes some unpredicted consequences.