Make all coercions not safe by default
This is the first step towards #291
Now all coercion types raise an error when failing.
I also made nominal type completely check-less. I'm not sure if this is the final decision. Probably, we'll need to extend the type interface to add sane behavior to nominal types, as in `try_coerce` or something like that.