Add wrapper type
It's a POC of the wrapper type. It's an analogue of rack middleware where one piece of middleware explicitly calls next one and then handles the result. Similarly, wrapper types form a call chain where you can explicitly modify both input and output. In a way, it's an extension of the constructor type.